% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{actualize_transformation}
\alias{actualize_transformation}
\title{Actualize a transform}
\usage{
actualize_transformation(obj)
}
\arguments{
\item{obj}{mesh3d object}
}
\value{
transformed object
}
\description{
Applying transformations to a mesh3d object only update the object's
transformation matrix, and does not actually modify any vertices.
}
\details{
This function applies the transformation to the vertices.
}
