% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-add-normals.R
\name{add_normals.mesh3d}
\alias{add_normals.mesh3d}
\title{Add face and vertex normals for mesh3d objects.}
\usage{
\method{add_normals}{mesh3d}(x, ...)
}
\arguments{
\item{x}{mesh3d object}

\item{...}{ignored}
}
\description{
This is a modified version of \code{rgl::addNormals()}. The \code{rgl} function
only saves the vertex nomarls - but it does actually calculate face normals
as an intermediate result.
}
\details{
This function is a clone of \code{rgl::addNormals()} where these face normals
are assigned to the object rather than just being discarded.

There is also some extra logic so that no attempt is made to calculate normals
of 'point' and 'line' objects.
}
