% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-add-normals.R
\name{add_normals}
\alias{add_normals}
\title{Add face and vertex normals to an object}
\usage{
add_normals(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other arguments}
}
\description{
Currently only 'mesh3d' objects are supported.
}
