% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{add_zorder}
\alias{add_zorder}
\title{Add zordering}
\usage{
add_zorder(obj_df)
}
\arguments{
\item{obj_df}{an object data.frame}
}
\value{
data.frame with 'zorder' column
}
\description{
Add zordering
}
