% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{add_zorder_multiple_objects}
\alias{add_zorder_multiple_objects}
\title{Add zordering to an object with multiple object_ids}
\usage{
add_zorder_multiple_objects(obj_df)
}
\arguments{
\item{obj_df}{an object data.frame}
}
\value{
data.frame with 'zorder' column
}
\description{
Add zordering to an object with multiple object_ids
}
