% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{as.data.frame.mesh3d}
\alias{as.data.frame.mesh3d}
\title{Convert mesh3d object to a data.frame representation}
\usage{
\method{as.data.frame}{mesh3d}(x, ...)
}
\arguments{
\item{x}{mesh3d object}

\item{...}{ignored}
}
\value{
data.frame representation
\itemize{
\item{\code{element_id}} - {element index}
\item{\code{element_type}} - {an integer from 1 to 4 indicating whether the element is
a point, line, triangle or quad}
\item{\code{vorder}} - {vertex ordering within element}
\item{\code{x,y,z}} - {coorindates of vertex}
\item{\code{vertex}} - {global index of this vertex}
\item{\code{vnx,vny,vnz}} - {unit vector in direction of vertex normal}
\item{\code{fnx,fny,fnz}} - {unit vector in direction of face normal}
\item{\code{fcx,fcy,fcz}} - {coordinates of centroid of element}
\item{\code{zorder}} - {\code{zorder} is used to control ggplot draw order.
                      For polygons it is based upon \code{fcz}. For lines it is
                      the smaller of the \code{z} coordinates. For points, it is the
                      \code{z} coordinate of the point}
\item{\code{zorder_var}} - { The value used to determine zorder }
\item{\code{hidden}} - by assuming the camera faces along the negative z-axis,
any triangle or quad face which has a face normal which also points along the
negative z-axis faces away from the camera and is hidden. Boolean value.

}
}
\description{
Convert mesh3d object to a data.frame representation
}
