% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{as.data.frame.mesh3dlist}
\alias{as.data.frame.mesh3dlist}
\title{Convert all objects in mesh3d list into a single data.frame}
\usage{
\method{as.data.frame}{mesh3dlist}(x, ...)
}
\arguments{
\item{x}{mesh3dlist object}

\item{...}{ignored}
}
\description{
An \code{object_id} will also be included in the output data.frame
}
