% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{as.mesh3d}
\alias{as.mesh3d}
\title{Convert the given object to a mesh3d object}
\usage{
as.mesh3d(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{arguments to particular methods}
}
\description{
Convert the given object to a mesh3d object
}
