% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot3D.r
\name{barplot3D}
\alias{barplot3D}
\title{Function barplot3D}
\usage{
barplot3D(
  x,
  col = NULL,
  transparency = 0.2,
  xlim = NULL,
  ylim = NULL,
  main = "3D Barplot",
  thetaX = 10,
  thetaY = 10,
  thetaZ = 0,
  add.numbers = F,
  txt.srt = 0,
  txt.pos = 3
)
}
\description{
Function barplot3D
}
