% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{create_bbox_obj}
\alias{create_bbox_obj}
\title{Create a bounding box for the given object at
the correct scale and location}
\usage{
create_bbox_obj(obj)
}
\arguments{
\item{obj}{3d object}
}
\value{
bounding box (mesh3d cube object) for the given 3d object
}
\description{
Create a bounding box for the given object at
the correct scale and location
}
