% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-conversion.R
\name{fortify.mesh3d}
\alias{fortify.mesh3d}
\title{Fortify a mesh3d object into a data.frame for compatibility with ggplot2}
\usage{
fortify.mesh3d(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{unused}
}
\value{
data.frame
}
\description{
Fortify a mesh3d object into a data.frame for compatibility with ggplot2
}
