% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive-types.R
\name{get_element_types}
\alias{get_element_types}
\alias{get_element_types.data.frame}
\alias{get_element_types.mesh3d}
\title{Determine element types present in object.}
\usage{
get_element_types(x)

\method{get_element_types}{data.frame}(x)

\method{get_element_types}{mesh3d}(x)
}
\arguments{
\item{x}{object (mesh3d object or data.frame of elements)}
}
\value{
Integer vector with values in range 1:4 representing the primitive types
        c('point', 'line', 'triangle' or 'quad')
}
\description{
Determine element types present in object.
}
