% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d-extract-matrix.R
\name{get_matrix_for_element_type}
\alias{get_matrix_for_element_type}
\title{Get a matrix of data to represent the given element type}
\usage{
get_matrix_for_element_type(obj, element_type, all_vertices = NULL)
}
\arguments{
\item{obj}{mesh3d object}

\item{element_type}{single element type to extract. specify an integer 1:4}

\item{all_vertices}{normalised matrix of all vertices. If NULL, then will be calculated within this call}
}
\value{
numeric matrix representation of the given element type
\itemize{
\item{\code{element_id}} - {element index}
\item{\code{element_type}} - {an integer from 1 to 4 indicating whether the element is
a point, line, triangle or quad}
\item{\code{vorder}} - {vertex ordering within element}
\item{\code{x,y,z}} - {coorindates of vertex}
\item{\code{vertex}} - {global index of this vertex}
\item{\code{vnx,vny,vnz}} - {unit vector in direction of vertex normal}
\item{\code{fnx,fny,fnz}} - {unit vector in direction of face normal}
\item{\code{fcx,fcy,fcz}} - {coordinates of centroid of element}
\item{\code{zorder}} - {\code{zorder} is used to control ggplot draw order.
                      For polygons it is based upon \code{fcz}. For lines it is
                      the smaller of the \code{z} coordinates. For points, it is the
                      \code{z} coordinate of the point}
\item{\code{zorder_var}} - { The value used to determine zorder }

}
}
\description{
`mesh3d` objects contain a matrix of vertices and matrix of vertex indicies
to describe each element.
}
\details{
This function maps the vertex indices to the actual vertex coordinates and
returns a matrix of information to represent each element of the requested
type.
}
