% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{get_object_info}
\alias{get_object_info}
\alias{get_object_info.mesh3d}
\title{Get a list of information about a 3d object}
\usage{
get_object_info(obj, ...)

\method{get_object_info}{mesh3d}(obj, ...)
}
\arguments{
\item{obj}{object}

\item{...}{arguments passed to particular methods}
}
\value{
list
\itemize{
\item{element_type} - {Type of element. point, line, triangle or quad}
\item{vertices} - {Number of vertices}
\item{elements} - {Number of elements}
\item{bytes} - {Size of object in bytes}
}
}
\description{
Get a list of information about a 3d object
}
