% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{identity_matrix}
\alias{identity_matrix}
\title{Create an identity matrix}
\usage{
identity_matrix(n = 4)
}
\arguments{
\item{n}{size. default: 4}
}
\value{
identity matrix of given size
}
\description{
Create an identity matrix
}
