% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivePolygonPlot.r
\name{interactivePolygonPlot}
\alias{interactivePolygonPlot}
\title{Function interactivePolygonPlot}
\usage{
interactivePolygonPlot(
  rver,
  pol,
  border = "blue",
  xlim = NULL,
  ylim = NULL,
  culling = "back"
)
}
\description{
Function interactivePolygonPlot
}
