% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interleave}
\alias{interleave}
\title{Interleave the elements of 2 vectors of the same length}
\usage{
interleave(a, b)
}
\arguments{
\item{a, b}{vectors}
}
\value{
interleaved values
}
\description{
Interleave the elements of 2 vectors of the same length
}
