% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-lookat.R
\name{look_at_matrix}
\alias{look_at_matrix}
\title{Create a camera 'look-at' transformation matrix i.e. camera-to-world transform}
\usage{
look_at_matrix(eye, at)
}
\arguments{
\item{eye}{position of camera (3-element vector)}

\item{at}{position in space to look at.  (3 element vector)}
}
\value{
homogenous 4x4 matrix
}
\description{
Create a camera 'look-at' transformation matrix i.e. camera-to-world transform
}
