% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3dlist.R
\name{mesh3dlist}
\alias{mesh3dlist}
\title{Create a list of mesh3d objects}
\usage{
mesh3dlist(...)
}
\arguments{
\item{...}{mesh3d objects}
}
\value{
an R object of type 'mesh3dlist'
}
\description{
Create a list of mesh3d objects
}
