% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-orthographic.R
\name{orthographic_projection}
\alias{orthographic_projection}
\title{orthographic projection}
\usage{
orthographic_projection(x, w = 2, h = 2, n = 1, f = 10)
}
\arguments{
\item{x}{matrix, mesh3d, vector or data.frame object}

\item{w}{width, height, near far frustrum boundaries}

\item{h}{width, height, near far frustrum boundaries}

\item{n}{width, height, near far frustrum boundaries}

\item{f}{width, height, near far frustrum boundaries}
}
\description{
Perform orthographic projection.
}
