% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-orthographic.R
\name{orthographic_projection_matrix}
\alias{orthographic_projection_matrix}
\title{Create a simple orthographic projection matrix}
\usage{
orthographic_projection_matrix(w = 2, h = 2, n = 1, f = 10)
}
\arguments{
\item{w, h, n, f}{width, height, near far frustrum boundaries}
}
\value{
orthographic projection matrix
}
\description{
Create a simple orthographic projection matrix
}
