% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-perspective.R
\name{perspective_projection}
\alias{perspective_projection}
\title{Perspective projection (symmetric frustrum)}
\usage{
perspective_projection(x, w = 2, h = 2, n = 1, f = 10)
}
\arguments{
\item{x}{matrix, mesh3d, vector or data.frame object}

\item{w}{width, height, near, far boundaries}

\item{h}{width, height, near, far boundaries}

\item{n}{width, height, near, far boundaries}

\item{f}{width, height, near, far boundaries}
}
\description{
The camera is at the origin looking along the negative-z axis.
}
