% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-perspective.R
\name{perspective_projection_matrix}
\alias{perspective_projection_matrix}
\title{Create a perspective projection matrix (symmetric frustrum)}
\usage{
perspective_projection_matrix(w = 2, h = 2, n = 1, f = 10)
}
\arguments{
\item{w, h, n, f}{width, height, near, far boundaries}
}
\value{
perspective projection matrix
}
\description{
Create a perspective projection matrix (symmetric frustrum)
}
