% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.r
\name{plotPoints}
\alias{plotPoints}
\title{Function plotPoints}
\usage{
plotPoints(
  tver,
  add = F,
  col = c("lightgrey", "black"),
  subset = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\description{
Function plotPoints
}
