% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolygons.r
\name{plotPolygons}
\alias{plotPolygons}
\title{Function plotPolygons}
\usage{
plotPolygons(
  rver,
  tver,
  pol,
  add = F,
  col1 = "lightgrey",
  col2 = "black",
  border = "black",
  culling = "none",
  lim = NULL,
  fun = mean
)
}
\description{
Function plotPolygons
}
