% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWireframe.r
\name{plotWireframe}
\alias{plotWireframe}
\title{Function plotWireframe}
\usage{
plotWireframe(
  tver,
  con,
  add = F,
  col = c("lightgrey", "black"),
  subset = NULL,
  xlim = NULL,
  ylim = NULL,
  fun = mean
)
}
\description{
Function plotWireframe
}
