% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh3d.R
\name{print.mesh3d}
\alias{print.mesh3d}
\title{Print mesh3d object}
\usage{
\method{print}{mesh3d}(x, max_cols = 10, ...)
}
\arguments{
\item{x}{mesh3d object}

\item{max_cols}{maximum number of columns to print for each matrix. default: 10}

\item{...}{ignored}
}
\description{
Print mesh3d object
}
