% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-rotate.R
\name{rotate_by}
\alias{rotate_by}
\title{Rotate the given object}
\usage{
rotate_by(x, angle, v)
}
\arguments{
\item{x}{matrix, mesh3d object or numeric vector of length 3 or 4 (homogenous coordinates)}

\item{angle}{Angle (radians)}

\item{v}{vector to rotate around}
}
\description{
Rotate the given object
}
