% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-rotate.R
\name{rotation_matrix}
\alias{rotation_matrix}
\title{Create a rotation matrix}
\usage{
rotation_matrix(angle, v)
}
\arguments{
\item{angle}{Angle (radians)}

\item{v}{vector to rotate around}
}
\value{
3d rotation matrix
}
\description{
from https://en.wikipedia.org/wiki/Rotation_matrix
}
