% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-scale.R
\name{scale_by}
\alias{scale_by}
\title{Scale the given object}
\usage{
scale_by(x, v)
}
\arguments{
\item{x}{matrix, mesh3d object or numeric vector of length 3 or 4 (homogenous coordinates)}

\item{v}{3 element vector giving scaling along each dimension}
}
\description{
Scale the given object
}
