% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-to-cube-at-origin.R
\name{scale_to_cube_at_origin.mesh3d}
\alias{scale_to_cube_at_origin.mesh3d}
\title{Resize to standard dimension i.e. within a 2x2x2 cube centred at origin}
\usage{
\method{scale_to_cube_at_origin}{mesh3d}(x, size = 2)
}
\arguments{
\item{x}{mesh3d object}

\item{size}{length of side of cube. default 2}
}
\description{
Resize to standard dimension i.e. within a 2x2x2 cube centred at origin
}
