% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-to-cube-at-origin.R
\name{scale_to_cube_at_origin}
\alias{scale_to_cube_at_origin}
\title{Resize to standard dimension i.e. within a 2x2x2 cube centred at origin}
\usage{
scale_to_cube_at_origin(x, size = 2)
}
\arguments{
\item{x}{matrix, mesh3d object or numeric vector of length 3 or 4 (homogenous coordinates)}

\item{size}{length of side of cube. default 2}
}
\description{
Resize to standard dimension i.e. within a 2x2x2 cube centred at origin
}
