% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-scale.R
\name{scaling_matrix}
\alias{scaling_matrix}
\title{Create a scaling matrix}
\usage{
scaling_matrix(v)
}
\arguments{
\item{v}{3 element vector giving scaling along each dimension}
}
\value{
3d scaling matrix
}
\description{
Create a scaling matrix
}
