% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-by.R
\name{transform_by}
\alias{transform_by}
\alias{transform_by.default}
\alias{transform_by.matrix}
\alias{transform_by.mesh3d}
\alias{transform_by.numeric}
\alias{transform_by.data.frame}
\alias{transform_by.mesh3dlist}
\alias{transform_by_inverse}
\title{Transform the vertex coordinates by the transformation matrix}
\usage{
transform_by(x, transform_matrix)

\method{transform_by}{default}(x, transform_matrix)

\method{transform_by}{matrix}(x, transform_matrix)

\method{transform_by}{mesh3d}(x, transform_matrix)

\method{transform_by}{numeric}(x, transform_matrix)

\method{transform_by}{data.frame}(x, transform_matrix)

\method{transform_by}{mesh3dlist}(x, transform_matrix)

transform_by_inverse(x, transform_matrix)
}
\arguments{
\item{x}{matrix, mesh3d object, mesh3dlist object, or numeric vector of length 3 or 4 (homogenous coordinates)}

\item{transform_matrix}{4x4 transformation matrix}
}
\description{
Transform the vertex coordinates by the transformation matrix
}
