% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformRver.r
\name{transformRver}
\alias{transformRver}
\title{Function transformRver}
\usage{
transformRver(
  rver,
  ver,
  thetaX = NULL,
  thetaY = NULL,
  thetaZ = NULL,
  persp = 0
)
}
\description{
Function transformRver
}
