% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-translate.R
\name{translate_by}
\alias{translate_by}
\title{Translate the given object}
\usage{
translate_by(x, v)
}
\arguments{
\item{x}{matrix, mesh3d object or numeric vector of length 3 or 4 (homogenous coordinates)}

\item{v}{3 element vector giving translation along each dimension}
}
\description{
Translate the given object
}
