% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-translate.R
\name{translation_matrix}
\alias{translation_matrix}
\title{Create a translation matrix}
\usage{
translation_matrix(v)
}
\arguments{
\item{v}{3 element vector giving translation along each dimension}
}
\value{
3d translation matrix
}
\description{
Create a translation matrix
}
