% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec3.R
\name{vec3_crossproduct}
\alias{vec3_crossproduct}
\title{Calculate the vector cross-product of 2 3d vectors}
\usage{
vec3_crossproduct(v1, v2)
}
\arguments{
\item{v1, v2}{vectors of length 3}
}
\value{
vector cross-product
}
\description{
Calculate the vector cross-product of 2 3d vectors
}
