% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec3.R
\name{vec3_dotproduct}
\alias{vec3_dotproduct}
\title{Calculate the vector dot-product of 2 3d vectors}
\usage{
vec3_dotproduct(v1, v2)
}
\arguments{
\item{v1, v2}{vectors of length 3}
}
\value{
vector dot-product
}
\description{
Calculate the vector dot-product of 2 3d vectors
}
