% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec3.R
\name{vec3_normalize}
\alias{vec3_normalize}
\title{Normalize a vector to be of unit length}
\usage{
vec3_normalize(v)
}
\arguments{
\item{v}{vector of length 3}
}
\value{
vector normalized to unit length
}
\description{
Normalize a vector to be of unit length
}
