% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.r
\name{plotPoints}
\alias{plotPoints}
\title{Function plotPoints}
\usage{
plotPoints(
  tver,
  add = F,
  col1 = "lightgrey",
  col2 = "black",
  subset = NULL,
  lim = NULL
)
}
\description{
Function plotPoints
}
