% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPolygons.r
\name{plotPolygons}
\alias{plotPolygons}
\title{Function plotPolygons}
\usage{
plotPolygons(
  tver = form("cube")$ver,
  pol = form("cube")$pol,
  add = F,
  col = "lightgrey",
  transparency = 0,
  border = "black",
  culling = "none",
  xlim = NULL,
  ylim = NULL,
  fun = mean
)
}
\description{
Function plotPolygons
}
