% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWireframe.r
\name{plotWireframe}
\alias{plotWireframe}
\title{Function plotWireframe}
\usage{
plotWireframe(
  rver,
  tver,
  con,
  add = F,
  col1 = "lightgrey",
  col2 = "black",
  subset = NULL,
  lim = NULL,
  fun = mean
)
}
\description{
Function plotWireframe
}
