% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scattergram3D.r
\name{scattergram3D}
\alias{scattergram3D}
\title{Function scattergram3D}
\usage{
scattergram3D(
  x,
  y,
  z,
  thetaX = 320,
  thetaY = 320,
  thetaZ = 0,
  col = c("darkgrey", "black"),
  border = NA,
  main = "3D Scattergram"
)
}
\description{
Function scattergram3D
}
