% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot3D.r
\name{surfaceplot3D}
\alias{surfaceplot3D}
\title{Function surfaceplot3D}
\usage{
surfaceplot3D(
  x = -10:10,
  y = -10:10,
  fun = NULL,
  dim = c(10, 10),
  border = NA,
  col = rgb(0, 0, 1),
  transparency = 0.1,
  thetaX = 10,
  thetaY = 60,
  thetaZ = 10,
  culling = "none",
  add.wireframe = T,
  add.polygons = T,
  ...
)
}
\description{
Function surfaceplot3D
}
